[UI] Custom branding for Cauldron Cloud
This change includes a new option in the configuration to change
the default name of Cauldron to a new name. This change is specially
focused on Cauldron Cloud instances that wants a custom name for their
instances.

CLOUD_CUSTOM_BRAND: name of the brand. (Cauldron)
CLOUD_CUSTOM_SUB_BRAND: 2 or 3 words describing the company. (Software Development Analytics)

Signed-off-by: JJMerchante <jj.merchante@gmail.com>